% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_mm.R
\name{power_mm}
\alias{power_mm}
\title{power_mm}
\usage{
power_mm(Formula, Varcomp, Resid_var = NULL, Data, Family, Effect, Alpha)
}
\arguments{
\item{Formula}{The formula to be passed into blmer.}

\item{Varcomp}{A vector of variance components to be fixed into the random
effects model. Values will fixed based on the order of random effects
specified in the formula statement.}

\item{Resid_var}{The residual variance value to be fixed. If the example does
not have a residual variance, type NULL.}

\item{Data}{The dataset to be used in blmer.}

\item{Family}{Specifies the distribution of the data.}

\item{Effect}{The random effect we are most interested in. This will be used
to calculate the power.}

\item{Alpha}{The critical value for the F-statistic. The default is 0.05.}
}
\description{
Calculates the power of a mixed effects model.
}
